requirements.txt
biopython
click
numpy
